Refactor and test sanitizeGraphqlName
