+++ src/utils.ts
@@ -252,18 +252,6 @@ export function getContentApiUrl(context: Context): string {
     return fixDoubleSlashes(`${API_URL}/${project}/${branch}`);
 }
 
-// sanitizes text according to graphql naming spec http://spec.graphql.org/October2021/#sec-Names
-export const sanitizeGraphqlName = (text: string) => {
-    if (!text || text.trim().length === 0) {
-        return '';
-    }
-    let result = text.replace(/([^0-9A-Za-z])+/g, '_');
-    if (result.length > 0 && /[0-9]/.test(result.charAt(0))) {
-        result = '_' + result;
-    }
-    return result;
-};
-
 // ---------------------------------------------------------------------------------------------------------------- Export
 
 const adapterConstants = {
