+++ src/views/BaseMacro.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import {MacroConfig, MacroData, MetaData} from '../guillotine/getMetaData';
 import {ComponentRegistry} from '../ComponentRegistry';
-import {RENDER_MODE, sanitizeGraphqlName} from '../utils';
+import {RENDER_MODE} from '../utils';
+import {sanitizeGraphqlName} from '../utils/sanitizeGraphqlName';
 import HTMLReactParser from 'html-react-parser';
 
 const unescape = require('unescape');
